made topics its own component
