+++ client/src/components/ChatApp.jsx
@@ -3,6 +3,7 @@ import { useEffect, useState } from "react";
 import io from "socket.io-client";
 
 import { Message } from "./Message";
+import { Topics } from "./Topics";
 
 const socket = io("http://localhost:5001");
 
@@ -41,26 +42,14 @@ export const ChatApp = () => {
     setNewMessage("");
   };
 
-  const handleJoin = (e) => {
-    socket.emit("leave topic", currentTopic);
-    setCurrentTopic(e.target.value);
-    socket.emit("join topic", e.target.value);
-  };
-
   return (
     <div>
-      <button value="1" onClick={handleJoin}>
-        Join topic 1
-      </button>
-      <button value="2" onClick={handleJoin}>
-        Join topic 2
-      </button>
-      <button value="3" onClick={handleJoin}>
-        Join topic 3
-      </button>
-      <button value="general" onClick={handleJoin}>
-        general
-      </button>
+      <Topics
+        socket={socket}
+        setCurrentTopic={setCurrentTopic}
+        currentTopic={currentTopic}
+      />
+
       <h1>Welcome To Topic: {currentTopic}</h1>
       <div>{listItems}</div>
       <form onSubmit={handleMessageSend}>
